refac: Line에서 Stations로 remove 로직 추출
